refactor: rename metrics.ts
